Removed dev tools component and integrated redux-devtools-extension
