+++ views/store/configureStore.dev.js
@@ -2,14 +2,14 @@ import { createStore, applyMiddleware, compose } from 'redux';
 import thunk from 'redux-thunk';
 import createLogger from 'redux-logger';
 import rootReducer from '../reducers';
-import DevTools from '../containers/DevTools';
 
 export default function configureStore() {
+  const composeEnhancers = window.__REDUX_DEVTOOLS_EXTENSION_COMPOSE__ || compose;
+
   const store = createStore(
     rootReducer,
-    compose(
-      applyMiddleware(thunk, createLogger()),
-      DevTools.instrument()
+    composeEnhancers(
+      applyMiddleware(thunk, createLogger())
     )
   );
